:art: Split svm implementation into multiple files
