+++ woke/woke/b_svm/exceptions.py
@@ -1,0 +1,4 @@
+class ChecksumError(Exception):
+    """
+    Checksum of a downloaded file did not match the expected value.
+    """
